+++ client/composables/useBooklet.ts
@@ -84,6 +84,14 @@ export default function useBooklet() {
     })
   }
 
+  async function regenerateDeletedPrevisionalTransactions(
+    bookletId: string,
+    month: number,
+    year: number,
+  ): Promise<TransactionResultDTO[]> {
+    return post(`booklet/${bookletId}/transactions/regenerate?month=${month}&year=${year}`, undefined)
+  }
+
   return {
     createBooklet,
     fetch: findAllBooklet,
@@ -93,5 +101,6 @@ export default function useBooklet() {
     findByIdMonthAndYear,
     findBalancesByIdMonthAndYear,
     findTransactionsByIdMonthAndYear,
+    regenerateDeletedPrevisionalTransactions,
   }
 }